FIxed Lecturer details update and delete functions

+++ TimetableManagementSystem/Lecturers/AddLecturer.cs
@@ -17,6 +17,7 @@ namespace TimetableManagementSystem.Lecturers
         public AddLecturer()
         {
             InitializeComponent();
+
         }
 
         SqlConnection con = Config.con;
@@ -93,41 +94,39 @@ namespace TimetableManagementSystem.Lecturers
 
                 GetLecturers();
 
-                ClearFields();
+                ClearFieldsAfterAdd();
 
                 tabControlLecturers.SelectedTab = tabPageLecView;
             }
         }
 
-        private void ClearFields()
+        private void ClearFieldsAfterAdd()
         {
             txtLecName.Clear();
-            txtLecDep.Clear();
             //cmbLecFac.SelectedIndex = -1;
-            cmbLecBuilding.SelectedIndex = -1;
+            cmbLecDepartment.SelectedIndex = -1;
             cmbLecCenter.SelectedIndex = -1;
+            cmbLecBuilding.SelectedIndex = -1;
             cmbLecLevel.SelectedIndex = -1;
             LecturerID = 0;
-            cmbLecDepartment.SelectedIndex = -1;
 
         }
 
         private bool IsValid()
         {
-            if ((txtLecName.Text == string.Empty) && (txtLecDep.Text == string.Empty) && (cmbLecFac.Text == string.Empty)
+            if ((txtLecName.Text == string.Empty) && (cmbLecFac.Text == string.Empty)
                 && (cmbLecCenter.Text == string.Empty) && (cmbLecBuilding.Text == string.Empty) && (cmbLecLevel.Text == string.Empty))
             {
                 MessageBox.Show("Fill the all fields", "Failed", MessageBoxButtons.OK, MessageBoxIcon.Error);
                 return false;
             }
 
-
             return true;
         }
 
         private void btnClear_Click(object sender, EventArgs e)
         {
-            ClearFields();
+            ClearFieldsAfterAdd();
         }
 
         private void dgvLectures_CellClick(object sender, DataGridViewCellEventArgs e)
@@ -135,13 +134,13 @@ namespace TimetableManagementSystem.Lecturers
             LecturerID = Convert.ToInt32(dgvLectures.SelectedRows[0].Cells[0].Value);
             txtLecNameEdit.Text = dgvLectures.SelectedRows[0].Cells[1].Value.ToString();
             cmbLecFacEdit.SelectedItem = dgvLectures.SelectedRows[0].Cells[2].Value;
-            //txtLecDepEdit.Text = dgvLectures.SelectedRows[0].Cells[3].Value.ToString();
             cmbLecDepartmentEdit.SelectedItem = dgvLectures.SelectedRows[0].Cells[3].Value;
             cmbLecCenterEdit.SelectedItem = dgvLectures.SelectedRows[0].Cells[4].Value;
             cmbLecBuildingEdit.SelectedItem = dgvLectures.SelectedRows[0].Cells[6].Value;
             cmbLecLevelEdit.SelectedItem = dgvLectures.SelectedRows[0].Cells[5].Value;
 
             tabControlLecturers.SelectedTab = tabPageLecEdit;
+
         }
 
         private void btnLecUpdate_Click(object sender, EventArgs e)
@@ -153,7 +152,7 @@ namespace TimetableManagementSystem.Lecturers
 
                 cmd.Parameters.AddWithValue("@LecName", txtLecNameEdit.Text);
                 cmd.Parameters.AddWithValue("@lecfaculty", cmbLecFacEdit.Text);
-                cmd.Parameters.AddWithValue("@lecdepartment", txtLecDepEdit.Text);
+                cmd.Parameters.AddWithValue("@lecdepartment", cmbLecDepartmentEdit.Text);
                 cmd.Parameters.AddWithValue("@leccenter", cmbLecCenterEdit.Text);
                 cmd.Parameters.AddWithValue("@lecbuilding", cmbLecBuildingEdit.Text);
                 cmd.Parameters.AddWithValue("@leclevel", cmbLecLevelEdit.Text);
@@ -193,44 +192,61 @@ namespace TimetableManagementSystem.Lecturers
                 cmd.ExecuteNonQuery();
                 con.Close();
 
-                MessageBox.Show("Lecturer details Updated", "Successfully", MessageBoxButtons.OK, MessageBoxIcon.Information);
+                MessageBox.Show("Lecturer Details Updated", "Successfully", MessageBoxButtons.OK, MessageBoxIcon.Information);
 
                 GetLecturers();
 
-                ClearFields();
+                ClearFieldsAfterUpdate();
 
                 tabControlLecturers.SelectedTab = tabPageLecView;
             }
             else
             {
-                MessageBox.Show("Please select a name to update ", "Select?", MessageBoxButtons.OK, MessageBoxIcon.Error);
+                MessageBox.Show("Please Select a lecturer to Update ", "Select?", MessageBoxButtons.OK, MessageBoxIcon.Error);
             }
+        }
+
+        private void ClearFieldsAfterUpdate()
+        {
+            txtLecNameEdit.Clear();
+            //cmbLecFac.SelectedIndex = -1;
+            cmbLecDepartmentEdit.SelectedIndex = -1;
+            cmbLecCenterEdit.SelectedIndex = -1;
+            cmbLecBuildingEdit.SelectedIndex = -1;
+            cmbLecLevelEdit.SelectedIndex = -1;
+            LecturerID = 0;
+
         }
 
         private void btnLecDelete_Click(object sender, EventArgs e)
         {
             if (LecturerID > 0)
             {
-                SqlCommand cmd = new SqlCommand("DELETE FROM Lecturers WHERE LecturerID = @LecturerID", con);
-                cmd.CommandType = CommandType.Text;
 
-                cmd.Parameters.AddWithValue("@LecturerID", this.LecturerID);
+                if(MessageBox.Show("Are You Sure You Want to Delete the Lecturer?","Delete Lecturer",MessageBoxButtons.YesNo,MessageBoxIcon.Information)==DialogResult.Yes)
+                {
+                    SqlCommand cmd = new SqlCommand("DELETE FROM Lecturers WHERE LecturerID = @LecturerID", con);
+                    cmd.CommandType = CommandType.Text;
 
-                con.Open();
-                cmd.ExecuteNonQuery();
-                con.Close();
+                    cmd.Parameters.AddWithValue("@LecturerID", this.LecturerID);
 
-                MessageBox.Show("Name is Deleted", "Successfully", MessageBoxButtons.OK, MessageBoxIcon.Information);
+                    con.Open();
+                    cmd.ExecuteNonQuery();
+                    con.Close();
 
-                GetLecturers();
+                    MessageBox.Show("Lecturer is Deleted", "Successfully", MessageBoxButtons.OK, MessageBoxIcon.Information);
 
-                ClearFields();
+                    GetLecturers();
 
-                tabControlLecturers.SelectedTab = tabPageLecView;
+                    ClearFieldsAfterUpdate();
+
+                    tabControlLecturers.SelectedTab = tabPageLecView;
+                }
+                
             }
             else
             {
-                MessageBox.Show("Please select a name to delete ", "Select?", MessageBoxButtons.OK, MessageBoxIcon.Error);
+                MessageBox.Show("Please Select a lecturer to Delete ", "Select?", MessageBoxButtons.OK, MessageBoxIcon.Error);
             }
         }
 
@@ -277,17 +293,31 @@ namespace TimetableManagementSystem.Lecturers
                 cmbLecDepartment.Items.Add("BM");
                
             }
-            else
-            {
-
-            }
 
             //cmbLecFac.SelectedIndex = -1;
         }
 
         private void cmbLecFacEdit_SelectedIndexChanged(object sender, EventArgs e)
         {
-            
+            cmbLecDepartmentEdit.Items.Clear();
+
+            if (cmbLecFacEdit.SelectedItem.Equals("Engineering"))
+            {
+                cmbLecDepartmentEdit.Items.Add("CE");
+                cmbLecDepartmentEdit.Items.Add("EE");
+            }
+            else if (cmbLecFacEdit.SelectedItem.Equals("Computing"))
+            {
+                cmbLecDepartmentEdit.Items.Add("CSSE");
+                cmbLecDepartmentEdit.Items.Add("DS");
+                cmbLecDepartmentEdit.Items.Add("IT");
+
+            }
+            else if (cmbLecFacEdit.SelectedItem.Equals("Business"))
+            {
+                cmbLecDepartmentEdit.Items.Add("IM");
+                cmbLecDepartmentEdit.Items.Add("BM");
+            }
         }
     }
 }
